Put Tree in its own module
This means only the code in tree.rs has the ability to violate the relevant invariants
